constructor error handling (#69)
* FfiResults on constructors, general error handling

removed unwraps and expects, added a unwrap_assert trait to differentiate unwraps in situations with unreachable errors

* push error handling down into ffi utils

* debug parsing column names from FfiObject

* don't panic upon failed dispatch

* cleanup/documentation

* post-rebase fixes

* enforce existence of .unwrap_assert explanations at compile-time

* fix fn documentation on ExplainUnwrap

* defer backtrace resolution until FfiResult::from

I was noticing a ~10 second lag in some cases. This is called by the symbol resolution phase of backtrace. This opens some interesting timing channels- we need to be careful about construction of errors, especially as our binary gets larger.

* return FfiResults from free functions

* fix tests

* tweak tests; add lint to error_free

the lint doesn't do much, since we're not calling it from rust

* fix tests

pay attention Mike