Interactive Measurements (and type-erased component types) (#89)
Closes #80.

I kept this as a fairly conservative change. Most of the code is in separate modules (any and interactive). Aside from a few trait bounds, the existing types and logic stay the same for non-interactive Measurements and Transformations.

If/when we're happy with this, we can migrate the FFI wrappers to use the type-erased (AnyXXX) approach in #88.

    Created type-erased versions of Measurement/Transformation component types.
    Added type-erasing transformers to Measurement/Transformation.
    Defined generic Queryable abstraction.
    Implemented make_adaptive_composition() constructor.
    Added unit tests.
    Created a separate abstraction for polymorphic operations.



* * Created type-erased versions of Measurement/Transformation component types.
* Added type-erasing transformers to Measurement/Transformation.
* Defined generic Queryable abstraction.
* Implemented make_adaptive_composition() constructor.
* Added unit tests.

* * Created a separate abstraction for polymorphic operations.