harden geometric mechanism
Adjusted the geometric mechanism to use the existence of bounds to run the algorithm in constant-time. When running in constant-time, the distribution is truncated to bound the worst-case runtime. When the distribution is truncated, the output needs to be clamped, because the tail probabilities are not simulated. Also adjusted the geometric sampler to take a shift and direction argument. This makes it possible to use the min and max values of T as the bounds when not running the algorithm in constant time-- if noise were represented in a signed data type we could not assume the min and max values as bounds of the data type, because a (positive int - noise) is greater than the min value on the type. Therefore since the geometric sampler now has a shift parameter, it is now possible to have a non-constant-time version of the geometric that doesn't need lower and upper bounds. Split the FFI into separate functions. Default is non-constant.