hardened geometric mechanism, data processing, merged constructors (#104)
* in-progress imputation

* add make_cast_vec_nullable

* make_impute_constant generic over representation of nullity

* rename NullDomainConstant to ImputeNull

naming goes a long way towards readability...

* update documentation

* experimental merged clamp constructor

this one make_ function can operate on both vectors and scalars

* further imputation polish

* check that constant is not null

* promote AllDomain<Option<T>>

Option<T> is a valid type. Using a specific domain allows to keep the distinction between null and valid types. Also restricted the carrier types of InternalNullDomain with an InternalNull trait.

* DatasetMetric inherits from Metric<Distance=u32>

explicitly setting the bit-depth in the inherit simplifies trait bounds throughout the library

* drop MO from clamp with associated type; fix ffi compilation

* rust ffi layer for impute, clamp, cast

* merge unclampers; IntervalDomain bounds checks

* synchronize rust ffi with opendp crate

this doesn't yet include the python bindings

* merge mechanisms; saturating and vector geometric

additions inside the geometric mechanism should be saturating, not default back to zero. Also added vector geometric mechanism

* const-generic LPSensitivity; update ffi/docs/tests

* harden geometric mechanism

Adjusted the geometric mechanism to use the existence of bounds to run the algorithm in constant-time. When running in constant-time, the distribution is truncated to bound the worst-case runtime. When the distribution is truncated, the output needs to be clamped, because the tail probabilities are not simulated. Also adjusted the geometric sampler to take a shift and direction argument. This makes it possible to use the min and max values of T as the bounds when not running the algorithm in constant time-- if noise were represented in a signed data type we could not assume the min and max values as bounds of the data type, because a (positive int - noise) is greater than the min value on the type. Therefore since the geometric sampler now has a shift parameter, it is now possible to have a non-constant-time version of the geometric that doesn't need lower and upper bounds. Split the FFI into separate functions. Default is non-constant.

* update geometric explanations, and clamp inputs

* update geometric documentation

* update geometric documentation

* resolve review comments

thanks Andy!

* add python tests