alp_histogram ffi
This adds a constructor for make_alp_parameterized on the python side that returns a measurement that emits a queryable when invoked. Implemented FFI support for Queryables, including AnyQueryables and state wrapping (to include Types), unpacking AnyObjects on the rust side into FfiSlice of a AnyQueryable, and the python-side unpacking of FfiSlices into AnyQueryables. Added lib functions for retrieving the input type of an AnyQueryable, invoking a queryable, and dropping a queryable, along with the python abstractions to make these transparent to use. Added traits for converting from Measurements that emit AllDomain<Queryable<S, Q, A>> into AnyMeasurements where the output AnyDomain's carrier converts Queryable<S, Q, A> into AnyQueryables first. Also adjusted the alp histogram to depend on the Hash trait into the ToPrimitive trait, so that non-numeric hashable keys are supported. This change is still subject to review from Christian.