replace relations with maps (#463)
* initial replacement of relations with maps

* drop delta maps

going forward, as we have multi-parameter distances, we'll want to be opinionated on the map directions. If we decide we want to keep delta maps later, we can add them. It is more difficult to remove them. Generally speaking, binary search should be suitable for solving for delta.

* make_fix_delta combinator

* fix make_population_amplification, conservative rounding

* analytic gaussian forward map, .check uses TotalOrd

* fix python tests

* fix rust doctests

* remove rust binary search utilities

* fix python doctests

* fix python notebooks

* add latex context for analytic gaussian

* remove spurious IntDistance: InfCast<_>

this is no longer needed for the backward maps!

* Use InfDiv when estimating amplification sampling rate

* make_fix_delta type introspection

adds an explicit import of core to the combinator module

* address review comments

thanks Andy!

* address review comments

thanks Andy!