QOL adjustments; python utility to specify default atomic type (#492)
* misc tweaks

rename fn make_resize_constant to fn make_resize to be consistent with python. In resize, move TA to the last type argument. Add a check to RuntimeType to ensure that the lists are homogeneous. Include type information in the Debug impl for domains, metrics and measures.

* move const P before distance Q in LpDistance<P, Q>

* set_default_int_type and set_default_float_type

* fix tests

* add docstring to utility functions

* address review comments

thanks Andy! Should have caught these myself!