grouping columns example (#1508)
- Fix #1488

This runs locally with faker and pandas installed, but I think we'd
prefer not to add those as dev dependencies?

I'm also just not sure how persuasive this example is: It feels like
we're doing a lot of work to just pull one example from a very small
set.

Proposal: The fake dataset should have a lot of columns, and the column
data should vary systematically, so we can see that it's picking a good
set of columns. My hope is that if we gave it
- 10 columns of constant values
- 10 columns of small random integers 0-9
- and 10 columns of random floats

it would pick the small random integer columns for grouping. Is that the
behavior we'd expect, and would it make a good example?