Upgrade pydata theme for search snippets (#1640)
- Fix #1517 

Confirmed locally that search now shows snippets. The only other
difference I see with this release is the styling of tabs: They seem
clunkier to me, but I don't think it's something we should fight.

<img width="499" alt="Screenshot 2024-05-21 at 5 43 41 PM"
src="https://github.com/opendp/opendp/assets/730388/d463590b-b881-45c2-9288-386b3035900d">

When there is a final release, we should certainly use it, but I'm not
seeing any problems with the RC, and the search snippets are a big win.

Took me longer than I'd like to admit to confirm that it works locally.
The upstream fix is a change to the document template (to restore a
`role="main"` attribute) so there are a number a cache layers to clear:
- Checkout the branch and `pip install -r requirements.txt`
- Clear the sphinx build cache with `rm -r build/html`
- Because of browser security rules, it doesn't work with just
`file:///`: Instead you'll need `python -m http.server` or similar.
- My browser was also caching old versions of page, so clear the browser
cache.