Polars: OnceFrame (#1602)
Currently in Polars, Measurements emit LazyFrames. LazyFrames can be
repeatedly executed, and each time it is executed, the privacy loss
increases (but is not tracked anywhere).

This PR instead returns a queryable with the LazyFrame in its internal
state. When the user wants to execute the LazyFrame, the queryable
executes the LazyFrame and then frees it, making it impossible to
execute again.

A queryable is used so that the user can control how the LazyFrame is
executed: either collect into memory, sink to a CSV, sink to Parquet, to
IPC or to JSON. Queryables may also share the query plan, or allow the
user to configure query plan optimizations.

In both the Python and Rust APIs, the Queryable emitted mimics a (very)
limited subset of the LazyFrame API. We can expand this API in the
future as needed.