Add: Extra info on error if Python polars version doesn't match Rust (#1663)
For users, this doesn't enforce strict version matches... but if there
is an error, and there is a mismatch, it includes more information,
saying exactly what version it saw and what version it expected. With
the current state, I think it could be frustrating to be a downstream
user, particularly if some other library includes opendp as a
dependency, but forgot to reference our `[polars]` extra. The user is
going to have some trouble figuring out what version _would_ fix the
error.