Polars: laplace thresholding for private key-sets (#1716)
In this setting, the key set is not public:

```python
context = dp.Context.compositor(
    data=pl.LazyFrame(schema={"A": pl.Int32, "B": pl.String}),
    privacy_unit=dp.unit_of(contributions=1),
    privacy_loss=dp.loss_of(epsilon=1.0, delta=1e-7),
    split_evenly_over=2,
    margins={
        ("B",): dp.Margin(max_partition_length=5),
    },
)
```

However, you can still release grouped queries on it:
```python
query = (
    context.query()
    .group_by("B")
    .agg(pl.len().dp.noise(), pl.col("A").fill_null(2).dp.sum((0, 3)))
)
```

This is because the library adds thresholding to the query based on the
noisy `pl.len()` to satisfy approximate-DP.

## ProductOrd change
This PR also changes the behavior of `ProductOrd` on tuples:
A tuple is only considered greater than another if all elements are
greater than the corresponding element in the other tuple. The library
previously used a lexicographic ordering, which, while this makes tuples
totally ordered, effectively only guarantees that the epsilon in the
epsilon/delta tuple is solved for. This is not really what we want from
the check in our maps: the check should only succeed if both epsilon and
delta are satisfied.

While the new product ordering makes check perform as expected, it also
means tuples no longer have a total ordering, so binary searches over
approx-DP mechanisms now fail. I've adjusted all of these call locations
to first solve for epsilon, and then delta.