Polars: no kwargs in public APIs (#1767)
Closes #1761

The design here is that each operation we want to add to Polars is
implemented with two plugins:
* a shim, that takes no keyword arguments. What would typically be
keyword arguments are passed as expressions.
* the actual plugin, which uses keyword args (and pickling). 

Users always write queries in terms of the shims. When OpenDP emits a
private lazyframe from the function, it swaps the shims out with the
real plugins.