report dir name if binary not found (#1886)
Proposal. Would help debug cases like 
- #1758 
- #1885
 
Hopefully we don't hit this error again, but if we did, we could ask the
user to list what was in that directory. It's also possible that someday
there will be a problem in the logic to determine the lib_dir, and this
would help us debug that.