Add required features note (#2014)
- Fix #1662
- Replace #1968
- Introduce pretty diffs in rust tests
- Add a test of docstring construction, but since it is only testing the
new code, it's pretty under-powered.

This also removes a bit of special logic for the "Features" section in
docstrings that was only used on one function. My sense is that keeping
both it and this new feature warning could be confusing. In the long
run, it would be good to explain places where `honest-but-curious` is
required, but we should make sure that everything is in sync.

- See #1664