Polars: simplify ExprDomain (#2157)
When writing proofs for Polars functionality, I found that there were
too many descriptors carried around in ExprDomain leading to more
complexity in proofs. Before this PR ExprDomain held a frame domain,
with information for all columns and margins, as well as information
about the context.

Since expressions always produce a single column, the refactored
ExprDomain only holds a single series domain instead of all series
domains.

When tracking margins in a row-by-row context, maintaining metadata
about margins that is consistent with all other margins is tricky, and
the descriptors quickly become extremely pessimistic to the point where
they are not useful. Due to this, I've completely removed access to
margin metadata when in the row-by-row context.

When tracking margins in a grouping context, only the margin about the
current grouping needs to be known. Therefore in the grouping context,
ExprDomains may hold one margin. This is useful for calibrating
sensitivity in aggregators.

By this logic, proof for output domain correctness need only consider
one series domain and how the margin might change in the grouping
context (which for most transformations will not change).

This PR also adds a `WildExprDomain` used to start expression
transformations/measurements when a column has not yet been
selected.`WildExprDomain` carries descriptors for all the select-able
series domains, and its carrier type is just a plan, which simplifies
how expression transformations/measurements are called (no more tuple of
a plan and `all()`)