Polars: simplify SeriesDomain (#2158)
SeriesDomain had a drawback where you could update the element domain
and forget to update the dtype, or vice-versa.

The refactored code retrieves the dtype from the element domain, and
eliminates the dtype descriptor. This also eliminates the use of the
Polars Field struct, shortening attribute access. I also added some
useful helpers for setting the dtype or element domain that better
reuses existing code.

The code also fixes a bug in the implementation of PartialEq for
SeriesDomain-- two series domains that differ in their nullity bit would
still be considered equal. This doesn't result in any potential exploits
though, since our idiom is to set the input space to the previous
domain's output space.