docs: Check for docstrings that are out of sync (#2181)
- Fix #1395

After looking for a library that would handle this, and coming back
empty-handed, decided just to wrap my own, and it didn't turn out to be
too complicated.

Can imagine expanding this to check that we do have the parameters in
the docstring, and that we have some description for each parameter, and
maybe check for return and typing info.

Going to branch from here and strengthen the tests, and start filling in
some gaps in the docs.