docs: fix PCA API docs (#2218)
- Fix #2217

Instead of making the base class dynamic (which causes trouble for
Sphinx), make it just a regular class, but update the class hierarchy if
the optional install is installed.

The core is getting rid of 
```python
_decomposition = import_optional_dependency('sklearn.decomposition', False)
if _decomposition is not None:
    class _SKLPCA(_decomposition.PCA): # type: ignore[name-defined]
        ...
else: # pragma: no cover
    class _SKLPCA(object): # type: ignore[no-redef]
        ...
```
and replacing it with
```
_decomposition = import_optional_dependency('sklearn.decomposition', False)
if _decomposition is not None:
    PCA = type('PCA', (_decomposition.PCA,), dict(PCA.__dict__))  # type: ignore[assignment,misc]
```

Since it's now documented by Sphinx, fill in docstring a little: could
still use more documentation on parameters, but this is a start.