feat: Example data in API (#2229)
- Fix #2216
- No shell commands in notebooks! More readable.
- Descriptions of examples can be found in our API docs
- More feasible to translate notebooks to RST.

I started with the helpers being StringIO, but I actually think now it
makes more sense to cache the actual files instead of relying on state
in the module. This way, each run should only require one download of
each CSV, and the user doesn't need to see the details.