docs: Checking docstrings, continued (#2182)
- Builds on #2181: Use that machinery to find things that can be
improved, and make the fixes.
- If there is a docstring, param tags must match signature.
- Public functions must have docstrings. (From just the AST, it is hard
to identify functions that are purely internal, so make many of them
explicitly private to exclude them from the check.)
- Check for unknown Sphinx directives.
- Check for consistency between type annotations and docs
- Check that directives are in usual order

Several blocks of code are commented out, with the number of warns that
arise if they are uncommented: I didn't want to introduce even more
changes that would make this hard to review and merge, but there is a
lot of follow-up work that could be done.

---------

Co-authored-by: Michael Shoemate <shoeboxam@gmail.com>