feat: Simplify Polars OO code (#2251)
- Inspired by #2218
- Fix #2231
- Replaces #2236 (new branch from main easier than sorting out
conflicts)

In this case it looks like we don't actually need anything from the
superclass, so we could just drop it, and we don't need to update the
type hierarchy or anything like this.

We do lose type annotations, because the types may not actually be
available, but those weren't showing up in the docs in any case, so I
don't think it's a big loss.