feat: Opendp object serialization (#2228)
- Replace #2192 
- Fix #2180
- Fix #2103 
- Follow-up issue: #2246

This takes a different approach to the serialization of OpenDP objects,
which at least at first seems like it might be simpler.

- Tweak the python.rs code gen to add a dictionary on objects.
- The original code had lists of the classes and modules to wrap. At
least for now I seem to be getting by without that.
- Calling the dictionary `__opendp_dict__` instead of `log`, just to
make it clear that this is special, and internal.
- Build on stdlib json serialization/deserialization, with the details
in private functions
- list vs. tuple: seems simpler to let lists flow straight through and
treat tuples as the special case.

---------

Co-authored-by: Michael Shoemate <shoeboxam@gmail.com>