docs: Check for doctrings on classes (#2260)
- Prompted by #2259

Make sure we document all of our public classes. The changes here do
that a few different ways.
- Fill in new documentation
- Move docstrings up from `__init__` that weren't being rendered. (This
is the default, but like everything else [it could be
configured](https://stackoverflow.com/questions/5599254/how-to-use-sphinxs-autodoc-to-document-a-classs-init-self-method).)
- Make classes that don't seem intended for end user use private. (Maybe
more of this?)

The `__init__` docs weren't being checked, but I could imagine checking
that the new class docstrings are consistent with constructor
parameters.