fix: replace _dependencies with lifelines (#2278)
Fix #2235

The plugin APIs embed Python objects inside opaque Rust structs
(transformation, measurement, queryable, function, domain, privacy
profile, etc).

Imagine a user defines a custom function and uses the plugin API to
embed it into a Rust struct.
When execution continues, the Rust struct holds a pointer to memory
allocated in Python.
Unfortunately, Python doesn't know the memory is still being referenced
by Rust,
and the Python refcount to that object naturally drops to zero.
The Python interpreter then frees the object upon the next garbage
collection, leading to a use-after-free memory corruption when the Rust
struct is used.

Previously, the generated code would add a reference to any Python
objects embedded in the Rust struct, to the corresponding opaque struct
on the Python side under the `_dependencies` attribute.
That is, the Rust object "depends on" the Python objects in its
`_dependencies`,
extending the lifetime of any embedded Python objects to the lifetime of
the Rust struct.
Two downsides to this approach:
- tracking the dependencies requires Rust structs built by combinators
to inherit dependencies of its arguments, leading to complexity
- there is no natural place to attach dependencies for privacy maps that
return user-defined Renyi-DP tradeoff functions

PR #2168 introduced a new approach to maintain the refcount, by adding a
"lifeline" to callback functions.
The lifeline is a struct that increments the Python refcount when
cloned, and decrements when dropped.
The intention of the PR was to capture a clone of the lifeline in the
Rust closure,
so that the Python refcount for the object would remain non-zero until
all Rust closures referencing the Python object/holding the lifeline are
dropped.
Unfortunately, the lifeline wasn't being captured, leaving the refcount
unprotected.

This PR ensures that the lifeline is captured, and adjusts transition
functions to use the same mechanism.
Since all Python objects are now protected via lifelines, this PR also
removes the dependencies system.



Disables R linting due to #2279