feat: `member` is true or false, with a warning; Expand serialization tests (#2265)
- Fix #1007
- Fix #2262

Python is less rigorous than Rust, and I think most users are just going
to expect `member` to return a bool. But we do have a side channel to
pass more information, if the user wants more information: warnings.

(This is motivated by wanting to write stronger tests of serialization,
but its not strictly necessary for that: I could just have `with
pytest.warns()` blocks.)