docs: quantiles theory section (#1734)
This is the quantiles notebook with polars and the breakdown of the
quantiles method using the transformation and measurements directly to
illustrate the intermediate steps.

unsure of where it would fit, but it can be split up. Incorporated
nearly all of Chuck's feedback from previous version and got a new
branch for better structure

- fix #1733

---------

Co-authored-by: Gurman Dhaliwal <gurmandhaliwal@Gurmans-MacBook-Air.local>
Co-authored-by: Chuck McCallum <mccalluc@users.noreply.github.com>
Co-authored-by: Chuck McCallum <mccallucc@gmail.com>
Co-authored-by: Shoeboxam <shoeboxam@gmail.com>