feat: Polars contribution bounding and sample/aggregate (#2318)
Close #1738

This PR grew to include general lib functionality needed for
contribution bounding:
* refactor to split distances in Margins out into a Bounds data type
* introduced the FrameDistance metric, whose distance type is a vector
of Bounds
* introduced dataset distances based on identifiers
* expanded FFI to allow for these new identifier metrics, as well as
bounded-DP metrics
* sample and aggregate framework, for infallible black-box functions
used in groupby truncations
* updated eurostat dataset to include user identifiers, and subsequently
tighten margin descriptors in examples

With this PR, three kinds of truncation may now be embedded in query
plans:
* per_group filter truncation
* per_group groupby truncation
* num_groups filter truncation

When parsing a query plan into a transformation, and the input metric is
an identifier distance, the library will first attempt to match
truncations. If it fails to match a truncation, the library will fall
back to the typical parser.