Make markdown explicit in docs (#2399)
- Fix #2332

Not a super-high priority, but I was trying to remember how this worked,
and simplifying it seemed like one way to do that.

- Instead of relying on heuristics, generated code has an explicit `..
end-markdown` between the markdown and the RST.
- We can use this flag to find the files that need processing, instead
of relying on special lists of files or location in hierarchy.
- Codegen is also responsible for generating RST parameter descriptions,
which I think just consists in doubling the backtics.
- Using a loop-break might be easier to read than the `next`?