docs: Check that all notebook cells are executed in order (#2388)
- Fix #2370

There's almost certainly a tool out there that already does this, but
didn't find it with a a quick search.

I think we're on the same page that this is a goal, but I don't think we
have a plan for how to get there. Would you like to just get all the
notebooks executed in this PR? Or is it something we don't even want to
worry about until release, and we don't want to worry about regressions
between releases? I would prefer something simple and consistent, but I
haven't been editing the notebooks as often as you.