docs: Remove orphan `noise` notebook (#2414)
The content here seems good, but it wasn't linked from any ToC. I can't
remember if there was a particular reason that we orphaned it? If it's
not ready to go live, then a reminder in the file itself, and a github
issue describing the work to be done might be useful.

- Fix #2408