docs: Explain constructor lingo (#2405)
- Fix #2364

I started looking at whether "constructor" could be replaced with
"function" across the board, but I don't think that makes sense: We
actually do mean something particular when we talk about constructors.
For instance, I don't think pure-python code in `extras` is considered
"constructors".

"Constructor" doesn't actually come up in `getting-started`, so this
seemed like a reasonable place, if it belongs anywhere.

Or we could just close this, if this doesn't seem like a real problem,
or the best fix to the problem.