feat: add `dp.sklearn.linear_model.LinearRegression` (#2403)
- fix #2395

We'll also want to update the documentation, probably keeping the
current example because it is a good example of a plugin, but adding a
pointer to this, if people just want to use it.

And I think we'll also want to provide access to this via polars? Maybe
the expression syntax will look like that offered by
https://github.com/azmyrajab/polars_ols?