docs: format code samples (#2448)
- Fix #2446
- Fix #2447 (Doesn't add a check, but I think things are standardized
now.)

Installed https://pypi.org/project/blacken-docs/ locally, and it seems
to work well. (Change from `python` to `pycon` required by
`blacken-docs`.)


Assuming it doesn't break any tests, this much seems harmless. Chose 59
as the linewidth for black because that seems to be the worst case with
the responsive docs site, until you make the window much narrower.

Regex to find some comment lines which are too long, though this isn't
everything.

If this much is useful, would it be good to do anything more here? We
decided against precommits earlier, as it adds one more step to get
started, but the line-length could be checked in CI, and the error there
could suggest running blacken-docs.