Clean up stats ToC in getting-started (#2494)
- Fix #2482
- Provide more explanation of the available tools: Mention sklearn
influence, and explain what synthetic data is good for... and not so
good for.
- Add a placeholder page for linear regression. Usually I want the
getting started to be self-contained, but we already have examples of
linear regression in two places, and I don't want to add a third, so
pointers will do. Having a statistics section and _not_ mentioning
linear regression feels like a gap.