Version changelog: v0.14.0 (#2504)
- Fix #2503

The instructions didn't work for me out of the box, so I've updated the
script and instructions: Not sure if this is the right direction.

One big question for me is what is `python channel_tool.py changelog`
supposed to do? It's a complicated script that has different behavior
depending on the environment, but it's not doing anything except adding
an empty line when I run it. I've dropped it from the instructions, but
not sure if that's best.

I'm going to open this PR, then do a self review and explain some of the
changes. Before merging this to main, I think it would make sense to
merge the couple last doc changes to main, and then merge main to this
branch.