Fix github CI syntax: only run the tests we need (#2506)
- Fix #2505
- We were running extra tests, I think, because string expressions just
evaluated to true
- But rather than using a lot of CI `if` checks, just put everything in
one case statement: The closer we can keep it to plain shell syntax, the
easier it will be to maintain.
- Use `==` instead of `contains`: Less likely to be surprised by a
partial match.
- Pull out just the name of the extra, and use a dummy `PLACEHOLDER`
when we don't want any extras: Less clutter.
- Add `numpy` to the test matrix: I'm not sure the extras need to be
this fine grained, but if we offer an extra, I think we're on the hook
for testing it.